updated generator versioning
